amend errors and move out of mod file
