+++ fil_fungible_token/src/token/mod.rs
@@ -2,111 +2,28 @@ use std::ops::{Neg, Rem};
 
 use cid::Cid;
 use fvm_ipld_blockstore::Blockstore;
-use fvm_ipld_encoding::Error as SerializationError;
 use fvm_ipld_encoding::RawBytes;
-use fvm_sdk::sys::ErrorNumber;
 use fvm_shared::address::Address;
-use fvm_shared::address::Error as AddressError;
 use fvm_shared::econ::TokenAmount;
 use fvm_shared::error::ExitCode;
 use fvm_shared::ActorID;
 use num_traits::Signed;
 use num_traits::Zero;
-use thiserror::Error;
 
-use self::state::StateInvariantError;
 use self::state::{StateError as TokenStateError, TokenState};
 use crate::receiver::types::TokenReceivedParams;
 use crate::runtime::messaging::{Messaging, MessagingError};
 use crate::runtime::messaging::{Result as MessagingResult, RECEIVER_HOOK_METHOD_NUM};
 use crate::token::TokenError::InvalidGranularity;
 
+mod error;
+pub use error::TokenError;
 pub mod state;
 pub mod types;
 
 /// Ratio of integral units to interpretation as standard token units, as given by FRC-XXXX.
 /// Aka "18 decimals".
 pub const TOKEN_PRECISION: u64 = 1_000_000_000_000_000_000;
-
-#[derive(Error, Debug)]
-pub enum TokenError {
-    #[error("error in underlying state {0}")]
-    TokenState(#[from] TokenStateError),
-    #[error("value {amount:?} for {name:?} must be non-negative")]
-    InvalidNegative { name: &'static str, amount: TokenAmount },
-    #[error("amount {amount:?} for {name:?} must be a multiple of {granularity:?}")]
-    InvalidGranularity { name: &'static str, amount: TokenAmount, granularity: u64 },
-    #[error("error calling receiver hook: {0}")]
-    Messaging(#[from] MessagingError),
-    #[error("receiver hook aborted when {operator:?} sent {amount:?} to {to:?} from {from:?} with exit code {exit_code:?}")]
-    ReceiverHook {
-        /// Whose balance is being debited
-        from: ActorID,
-        /// Whose balance is being credited
-        to: ActorID,
-        /// Who initiated the transfer of funds
-        operator: ActorID,
-        amount: TokenAmount,
-        exit_code: ExitCode,
-    },
-    #[error("expected {address:?} to be a resolvable id address but threw {source:?} when attempting to resolve")]
-    InvalidIdAddress {
-        address: Address,
-        #[source]
-        source: AddressError,
-    },
-    #[error("error during serialization {0}")]
-    Serialization(#[from] SerializationError),
-    #[error("error in state invariants {0}")]
-    StateInvariant(#[from] StateInvariantError),
-}
-
-impl From<TokenError> for ExitCode {
-    fn from(error: TokenError) -> Self {
-        match error {
-            TokenError::ReceiverHook { from: _, to: _, operator: _, amount: _, exit_code } => {
-                exit_code
-            }
-            TokenError::InvalidIdAddress { address: _, source: _ } => ExitCode::USR_NOT_FOUND,
-            TokenError::Serialization(_) => ExitCode::USR_SERIALIZATION,
-            TokenError::InvalidNegative(_) => ExitCode::USR_ILLEGAL_ARGUMENT,
-            TokenError::State(state_error) => match state_error {
-                StateError::IpldHamt(_) | StateError::Serialization(_) => {
-                    ExitCode::USR_SERIALIZATION
-                }
-                StateError::MissingState(_) => ExitCode::USR_ILLEGAL_STATE,
-                StateError::NegativeBalance { balance: _, delta: _, owner: _ }
-                | StateError::InsufficentAllowance {
-                    owner: _,
-                    operator: _,
-                    allowance: _,
-                    delta: _,
-                } => ExitCode::USR_FORBIDDEN,
-                StateError::NegativeTotalSupply { supply: _, delta: _ } => ExitCode::USR_FORBIDDEN,
-            },
-            TokenError::Messaging(messaging_error) => match messaging_error {
-                MessagingError::Syscall(e) => match e {
-                    ErrorNumber::IllegalArgument => ExitCode::USR_ILLEGAL_ARGUMENT,
-                    ErrorNumber::Forbidden | ErrorNumber::IllegalOperation => {
-                        ExitCode::USR_FORBIDDEN
-                    }
-                    ErrorNumber::AssertionFailed => ExitCode::USR_ASSERTION_FAILED,
-                    ErrorNumber::InsufficientFunds => ExitCode::USR_INSUFFICIENT_FUNDS,
-                    ErrorNumber::IllegalCid
-                    | ErrorNumber::NotFound
-                    | ErrorNumber::InvalidHandle => ExitCode::USR_NOT_FOUND,
-                    ErrorNumber::Serialization | ErrorNumber::IllegalCodec => {
-                        ExitCode::USR_SERIALIZATION
-                    }
-                    ErrorNumber::LimitExceeded => ExitCode::USR_UNSPECIFIED,
-                    _ => unreachable!(),
-                },
-                MessagingError::AddressNotInitialized(_) => ExitCode::USR_NOT_FOUND,
-                MessagingError::Ipld(_) => ExitCode::USR_SERIALIZATION,
-            },
-        }
-    }
-}
 
 type Result<T> = std::result::Result<T, TokenError>;
